Redigera som admin, visar endast tider framåt. favicon.

+++ Areas/Identity/Pages/Account/Manage/Index.cshtml.cs
@@ -1,10 +1,5 @@
 
-#nullable disable
 
-using System;
-using System.ComponentModel.DataAnnotations;
-using System.Text.Encodings.Web;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
@@ -12,6 +7,7 @@ using PadelApp.Data;
 using PadelserviceApp.Models;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace PadelApp.Areas.Identity.Pages.Account.Manage
 {
@@ -22,23 +18,29 @@ namespace PadelApp.Areas.Identity.Pages.Account.Manage
         private readonly SignInManager<IdentityUser> _signInManager;
 
         private readonly ApplicationDbContext _context;
+        public List<PadelBooking> Bookings { get; set; } = new List<PadelBooking>();
 
+        //private readonly ILogger<IndexModel> _logger;
+         public PadelBooking? PadelBooking { get; set; }
         public IndexModel(
             UserManager<IdentityUser> userManager,
             SignInManager<IdentityUser> signInManager,
             ApplicationDbContext context)
+            
         {
             _userManager = userManager;
             _signInManager = signInManager;
             _context = context;
+             PadelBooking = new PadelBooking();
+             
         }
 
    
-  //-------------------------Get--------------------------------------------//
+  //-------------------------Get(båda)--------------------------------------------//
 
-        public List<PadelBooking> Bookings { get; set; }
         public bool AdminCheck { get; set; }
 
+         DateTime sortDate = DateTime.Today;
      
         public async Task<IActionResult> OnGetAsync()
         {
@@ -50,7 +52,8 @@ namespace PadelApp.Areas.Identity.Pages.Account.Manage
         
         AdminCheck = await _userManager.IsInRoleAsync(user, "Admin");
         if (AdminCheck){
-        Bookings = await _context.PadelBookings
+        Bookings = await _context.PadelBookings //bara tider från idag och framåt
+            .Where(b => b.BookingTime >= sortDate) 
             .Include(b => b.Court)  
             .Include(b => b.User)   
             .ToListAsync();
@@ -58,7 +61,7 @@ namespace PadelApp.Areas.Identity.Pages.Account.Manage
             }
         if(!AdminCheck) {
             Bookings = await _context.PadelBookings
-                .Where(b => b.UserId == user.Id)
+                .Where(b => b.UserId == user.Id  && b.BookingTime >= sortDate) //bara tider från idag och framåt och samma användare
                 .Include(b => b.Court)  
                 .Include(b => b.User)   
                 .ToListAsync();
@@ -67,7 +70,7 @@ namespace PadelApp.Areas.Identity.Pages.Account.Manage
             return Page();
         }
 
-  //-------------------------delete--------------------------------------------//
+  //-------------------------delete(båda)--------------------------------------------//
         public async Task<IActionResult> OnPostDeleteAsync(int? id){
 
         var user = await _userManager.GetUserAsync(User);
@@ -81,15 +84,25 @@ namespace PadelApp.Areas.Identity.Pages.Account.Manage
             var Booking = await _context.PadelBookings
             .Include(p => p.User)
             .FirstOrDefaultAsync(m => m.BookingId == id);
+
+            if(Booking == null) {
+                return NotFound();
+            }
             _context.PadelBookings.Remove(Booking);
             await _context.SaveChangesAsync();
             return RedirectToPage();
         }
+
+
         if(!AdminCheck){
             var Booking = await _context.PadelBookings
             .Where(b => b.UserId == user.Id)
             .Include(p => p.User)
             .FirstOrDefaultAsync(m => m.BookingId == id);
+
+            if(Booking == null) {
+                return NotFound();
+            }
             _context.PadelBookings.Remove(Booking);
             await _context.SaveChangesAsync();
         }
